backend/api/requirements.txt
# Web framework
fastapi>=0.104.0
uvicorn>=0.24.0

# Database
sqlalchemy>=2.0.23
asyncpg>=0.29.0

# Object storage
aioboto3>=12.0.0
botocore>=1.31.0

# OCR and image processing
pytesseract>=3.10.1
Pillow>=10.1.0

# Document processing
PyPDF2>=3.0.1
python-docx>=1.1.0
openpyxl>=3.1.2

# Validation and settings
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Task scheduling
APScheduler>=3.10.4

# Environment and configuration
python-dotenv>=1.0.0